refactor(view): remove unused previous chart modules
