--- packages/view/src/components/TemporalFilter/ClocLineChart/ClocLineChart.const.ts
@@ -1,11 +1,0 @@
-export const WIDTH = 600;
-export const CLOC_STYLING = {
-  width: WIDTH,
-  height: WIDTH * 0.9,
-  padding: {
-    top: 0.1,
-    right: 0.3,
-    bottom: 0.3,
-    left: 0.1,
-  },
-};
