--- packages/view/src/components/TemporalFilter/ClocLineChart/ClocLineChart.scss
@@ -1,15 +1,0 @@
-@import "styles/_pallete";
-
-.cloc-line-chart-wrap {
-  height: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 20px;
-  margin: auto 0;
-}
-
-.cloc-line-chart {
-  overflow: visible;
-  fill: #ff8272;
-}
