--- packages/view/src/components/TemporalFilter/ClocLineChart/index.ts
@@ -1,1 +1,0 @@
-export { default as ClocLineChart } from "./ClocLineChart";
