--- packages/view/src/components/TemporalFilter/CommitLineChart/CommitLineChart.scss
@@ -1,15 +1,0 @@
-@import "styles/_pallete";
-
-.commit-line-chart-wrap {
-  height: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 20px;
-  margin: auto 0;
-}
-
-.commit-line-chart {
-  overflow: visible;
-  fill: #ff8272;
-}
