--- packages/view/src/components/TemporalFilter/CommitLineChart/index.ts
@@ -1,1 +1,0 @@
-export { default as CommitLineChart } from "./CommitLineChart";
